Add files via upload
This project is an implementation of graph data structures and algorithms, including the PageRank algorithm and Dijkstra's shortest path algorithm. The main part of the project is a C program for managing graphs, allowing the creation of graphs, adding vertices and edges, and computing and displaying the results of graph algorithms like PageRank and shortest paths.